+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * When aligning forms with `clojure-align` (or with the automatic align feature), blank lines will divide alignment regions.
 * [#378](https://github.com/clojure-emacs/clojure-mode/issues/378): Font-lock escape characters in strings.
+* Port threading macros related features from clj-refactor.el. Available refactorings: thread, unwind, thread first all, thread last all, unwind all.
 
 ## 5.3.0 (2016-04-04)